requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
importlib-metadata==3.4.0
Markdown==3.3.3
mysqlclient==2.0.3
psycopg2==2.8.6
psycopg2-binary
PyMySQL==1.0.2
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
vue==0.0.1
zipp==3.4.0
gunicorn==20.1.0 